server: Do not re-open NamedTemporaryFile on Windows
From the Python docs: "Whether the name can be used to open the file a
second time, while the named temporary file is still open, varies across
platforms (it can be so used on Unix; it cannot on Windows)"

The file-like object returned by NamedTemporaryFile can be used to write
to the temporary file. The writes must be flushed before another process
attempts to read the file.

Also, using an ExitStack removes the need to manually manage the
clean-up of the (optional) TemporaryDirectory and NamedTemporaryFile
used by `run_strictdoc_server()`.

Closes #1061