document_iterator: refactor the iteration algorithm
The previously developed concept of `ng_level` is not valid anymore because StrictDoc has to support exporting includable documents. If one document includes another one, they both have different level hierarchies, and the same node, e.g., a section or a requirement, can appear under different levels, such as 1.2.3 and 4.1.2.3.

This changeset moves the generation of levels from the .sdoc parsing phase to the document iteration phase. This also allow us to remove some of the oldest StrictDoc code, simplifying the overall implementation.