Feature: LINK to requirements and custom nodes
Currently, [LINK: ID] is limited to target either sections or anchors.

This change extends the feature so that we can link to REQUIREMENTs or
any other element from a custom grammar. E.g. the following construct
will work now:

[FREETEXT]
[LINK: CUSTOM-1]
[/FREETEXT]

[CUSTOM]
UID: CUSTOM-1

For cleaner design, we introduce a small wrapper class LinkableNode that
centralizes rules for how to create the link title.

To handle the case where a linked requirement is removed, we introduce
validation in core.transforms.