Remove Graphviz and DocumentDotGenerator
The experience has shown that it is not much that one can do with SVG or PDF files produced by graphviz. Visualizing requirements is easier and more practical with HTML and JavaScript.

If it happens that there is still any opportunity for using Graphviz with StrictDoc, this changeset should be easy to find and use it to revert this removal.