LINKs: Fix link to anchor in other sdoc
This kind of linking worked for [FREETEXT] (was covered by tests), but
not for [TEXT] (was not covered by tests) nodes.

Root cause was that Anchor.parent is an SDocNodeField which has no
parent_or_including_document. That property is defined on the
SDocNodeField.parent SDocNode. So we need Anchor.parent.parent.

Fixes #1905.