rst: Fix link followed by non-ws/punctuation
When rendering rst fragments to HTML, InlineLinks are first converted to
':rawhtml:`<a href="#TGT">link text</a>`', then are rendered to HTML by
docutils publish_parts. That rawhtml construct is a special form of
inline markup, for which rst defines recognition rules [1]. #7 says

  Inline markup end-strings must end a text block or be immediately
  followed by whitespace, one of the ASCII characters - . , : ; ! ? \ /
  ' " ) ] } > or a similar non-ASCII punctuation character. [20]

That's why LINK followed by character produces error or odd HTML.
To get around, we can use rst's escaping mechanism [2].

  Escaped non-white characters are prevented from playing a role in any
  markup interpretation. The escaped character represents the character
  itself.

The code change detects if an inline links is followed by non-whitespace
or non-punctuation character and adds in a backslash if so.

Punctuation characters in the non-latin range are not explicitly
checked and thus also joined with "\". This is unnecessary from RST
syntax point of view, but doesn't hurt as it just maps them to
themselves.

Fixes #1907.

[1] https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#inline-markup-recognition-rules
[2] https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#escaping-mechanism