links: Consitent anchor id whitespace replacement
Whitespaces are valid in StrictDoc UIDs, but are not valid in HTML
element ids. Therefore we have to sanitize them during HTML export. This
was already done in most places, but not for anchor IDs in rst
fragments. In consequence referencing links couldn't find their target.

This adds whitespace to dash replacement for rst-for-html anchors. It
reuses LinkRenderer for consistent results.

Fixes #1916.