Mark more special HTML in templates as safe
There were some occurrences of "&nbsp;" in Python statements within
Jinja2 templates. They must be marked as safe, otherwise they would be
rendered to literal &nbsp; by auto escaping.

This happened in diff screen for any text node, and sections with
LEVEL: None, as well as in TOCs for sections with LEVEL: None.

Relates to #1920.