backend/sdoc_source_code: enable scope=file marker for Python
This change also removes the coupling between the forward and backward links between requirements and source code. Previously, the backward links would only work from a source file to a requirement if there was a forward link from the requirement to the file. This is no longer needed and should have been done long time ago.