Skip binaries in source tree rather than failing
We're expecting source files to be UTF-8 text. Files that are invalid
UTF-8 (most likely binary files) would have crashed StrictDoc. We can
instead just skip them. Whatever they are, they are not traceable.

It's as simple as catching the UnicodeDecodeError. The majority of the
diff is just about remembering whether a file was successfully parsed
and indexed. Successfully parsed files are now in FileTraceabilityIndex

 TracabilityIndex
   has 1 FileTraceabilityIndex
     has n SourceFileTraceabilityInfo
       has 1 SourceFile  # new: remember parsed/indexed source file

while any found file is (as before) in DocumentTree

 Traceabilityindex
   has 1 DocumentTree
     has 1 SourceTree
       has n SourceFile  # remember *any* found source file

The binary files for testing was generated with
  echo "00 08 00 00" | xxd -r -p > invalid_file.bin