Represent forward file reference as marker
strictdoc handled the simple forward file relation listed below special:

[REQUIREMENT]
UID: REQ-001
TITLE: Requirement Title 1
STATEMENT: >>>
Requirement with forward relations.
<<<
RELATIONS:
- TYPE: File
  VALUE: file.py

All other file references (to functions, classes, lines, ranges, file
backwards) are represented as markers internally. Only the plain forward
file reference was not a marker, but an entry to a separate dict.

Let's resolve the simple forward file relation to a range marker
covering the whole file (1st line to last line) and get rid of the
special handling. This is consistent with what we get when we use a
@relation(..., scope=file) backward marker.