export/html: --included-documents: display or not display consistently
When the option was not provided, the included documents were not generated as expected. However, the links to the included documents were still generated on the project index screen. This changeset fixes the issue and additionally improves and centralizes handling of the conditions whether to show the included documents or not.