html export: minor review edits, adapt the Stimulus-based stable link button
The main action of this changeset is to adapt the Stimulus version of the stable link controller that we introduced here: https://github.com/strictdoc-project/strictdoc/pull/2158.

The rest are minor cleanups. In particular the itest #5 from the stable URL group was merged to the itest #1 because they both exercise an identical test setup.