Code climate: tree_cycle_detector and requirement_form_object: fix all mypy issues
This PR finally fixes the missing/incorrect cycle detector types.

This is triggered by the performance investigation around how the DTR screen generates Jinja templates for very large documents.

At some point, I suspected a bug in the logic because the type annotations were off, but it turned out that the logic was fine.