backend/sdoc_source_code: enable parsing of source nodes if a config is provided
This is a more permanent and proper solution for a use case where a user does not want to SDoc nodes to be parsed out of the source code comments. Parsing of SDoc nodes will from now on be only activated if the `source_nodes` config is provided in the `strictdoc.toml` config file.

Related to: #2342