screens/search: fix an incorrectly escaped search example
After the 2024-Q4 changes related to escaping of Jinja templates, escaping URLs by LinkRenderer can cause over-escaping if not wrapped into Markup(...). This only affected the example query found in the legend, and this commit fixes the issue. An e2e test is added to prove that the URL is rendered correctly.