docs: update the hello world example with a more realistic command
This will hopefully help the users to not stumble upon the edge cases like the one encountered by a user in the issue closed by this commit, see https://github.com/strictdoc-project/strictdoc/issues/2394#issuecomment-3110576718.

In that case, the user was running the "Hello, World!" example by providing a single file as input, rather than the containing folder. With this setup, StrictDoc could not discover the .sgra file located in the same folder as the document, because it was only processing the specified file and not considering the full contents of the folder.

Closes #2394