fix(parallelizer): switch to a more simple ProcessPoolExecutor API
With this, there is some hope that the multiprocessing edge cases on Windows and macOS will finally disappear.

Some years ago, I was testing this API with Python 3.5 and 3.6, and I could easily reproduce cases where the child processes became zombie processes. One easy case of getting zombies was to interrupt the integration tests. Testing this now, I could not reproduce anything to go wrong. Hoping that introducing this new API will not introduce its own new issues with flaky tests on GitHub.