feat(backend/sdoc_source_code): Support merging sdoc nodes with source nodes
This allows to have a (maybe incomplete) definition of a requirement
anywhere in the static document tree which is then merged with content
from annotated source code at runtime. Previously, StrictDoc would
always have auto-generated a new node in an auto-determined document
position. Match criterion is currently the UID, this should be enhanced
to MID in an upcoming change. If the grammar element of sdoc and source
code node are different, an error will be raised.

There are two use cases:

- Control the position (document, section) of a generated source node,
  as opposed to letting the node generator use an auto generated section.
- Support "side-car" files where parts of a requirement are defined in
  *.sdoc, while other parts of the same requirement are inlined with
  source code. This is a concept from the Linux kernel requirement
  template proposal.