fix(export/html): Consider MARKUP = Text option in source file view
Previously, RST was hard-coded.

Fix it by checking on a per-node basis if RST or Text rendering shall be
used. It's not possible to do this on a per-source-file basis, since
requirements for one source file may come from multiple documents and
the MARKUP option is attached to that documents, not to the source file.

Fixes #2544.