chore(backend/sdoc_source_code): MarkerParser class: require explicit arguments for better readability
The upcoming PR https://github.com/strictdoc-project/strictdoc/pull/2555 adds some more variables to this class, and it feels like it is time to enforce the explicit argument names every time this class's constructor is called. This should help with avoiding the issues when a wrong position argument can be set in place of the right one.