refactor(backend/sdoc_source_code): add test helpers for marker_lexer tests, add a MarkerParser-level Linux/SPDX test
This change is one more fragment from a larger https://github.com/strictdoc-project/strictdoc/pull/2555 extracted for an easier review.

The custom assertion helpers are introduced to the test_marker_lexer which reduces the test boilerplate.

The marker parser-level test_80_linux_spdx_example is introduced to verify that the comments are parsed correctly. Previously this test only existed at the marker lexer level.