feat(export/html): add ability to render git information in metadata fields for html and html2pdf
WHAT:

Add ability to render @GIT_VERSION, @GIT_BRANCH, @GIT_COMMIT_DATE, @GIT_COMMIT_DATETIME
not only in the reserved VERSION and DATE fields but also in user-defined metadata.

WHY:

Sometimes the users want to render Git information as part of their custom
document metadata.

Closes #2587

HOW:

Extend the DocumentScreenViewObject to support a method that preprocesses a
metadata value before rendering it to HTML or PDF. If the value contains one
of the known @GIT_* macros, the value is expanded to the corresponding value
obtained from the underlying Git information.