feat(server, html2pdf): implement better error reporting when Chrome is not found
**WHAT:**

This change improves the error message when the StrictDoc web server
cannot find a Chrome installation on the system it is running on.

**WHY:**

Before this change, the error message was a simple Internal Server Error
without an explicit indication that the Chrome could not be found. A
user would have to dig through the server logs to understand the issue.

**HOW:**

This integrates the upstream html2pdf4doc work where the error reporting
was improved in the main.py driver program.

In particular, the driver now exits with a specific exit code
`COULD_NOT_FIND_CHROME = 5` that we can rely on in StrictDoc to identify
the missing Chrome issue.

https://github.com/strictdoc-project/html2pdf4doc_python/issues/58
https://github.com/strictdoc-project/html2pdf4doc_python/pull/70