fix(link_renderer): fix links between two included documents
WHAT:

This change fixes a regression introduced by this commit:
https://github.com/strictdoc-project/strictdoc/commit/ec4d95c6b52fbf38fc4550432fbe2e3658490489.

WHY:

A bug report by a user who discovered the regression: #2595.

HOW:

With this fix, the case of one included document node linking to another
included document node is now handled by not letting the code to enter the
local-only path with `return f"#{local_link}"`. If the link renderer now sees
this case, it will not return early but generate a proper full relative path
to the target document node.

Closes #2595