fix(html2pdf): prevents the section titles staying as last elements on a page
WHAT:

This sets up the html2pdf4doc rule for making section titles go to the next
page if they have child nodes. In practice, this means that sections will
always jump to the next page if they have meaningful content.

WHY:

Closes #2450

HOW:

This is implemented by adding the `.html2pdf4doc-no-hanging` class that signals
to html2pdf4doc that this node cannot be the last element on a page and has to
go to the next page.