feat(html2pdf): introduce a new "html2pdf_forced_page_break_nodes" option
WHAT:

Introduce a new option that forces every node to start from a new page.

WHY:

This feature is especially useful for documents that contain very large,
descriptive requirement nodes with a lot of text. In such cases, starting each
node on a new page helps avoid having the beginning of a node lost in the middle
of a page.

This implements what a user asked for as part of this GitHub discussion:
https://github.com/strictdoc-project/strictdoc/discussions/2593

HOW: -