refactor(export/html): get html2pdf classes from view object, not Jinja
WHAT:

This extracts the logic for building html2pdf4doc classes into DocumentViewObject.

Also, we found an opportunity to remove the setting of html2pdf4doc classes
in the template: `components/node_content/index.jinja` because that logic is
not needed anymore.

WHY:

We want to avoid programming logic and control flow in Jinja. Python does this
much better.

We are planning to add more logic for customizing the html2pdf4doc control
classes, and doing this in Python will be much easier.

HOW:

- All Jinja logic has been moved to Python.