fix(source_files_finder): ignore all binary files at the first search step
WHAT:

This prevents StrictDoc from even trying to parse any binary files because
they are ignored at the earliest FileFinder search step.

WHY:

This resolves a bug report where a user reported that StrictDoc was crashing
while trying to read a binary file.

Closes #2582

HOW:

- The code to filter out binary files existed before, but the skipping of
  binary files was done too late. This caused a collision with the project
  statistics feature, which calculates source-code-related statistics on ALL
  files found by FileFinder. The binary files were included, causing the crash
  described in the bug report.
- Instead of fixing the issue at the project statistics level, the skipping of
  binary files was moved to the earliest search stage, so that StrictDoc never
  has to deal with binary files.
- A TODO for a future development is left based on the user suggestion:

```
TODO: For now, ignore the binary files but one day a user
might want to create a Relation to a binary file like a
published PDF.
```