feat(pyproject.toml): html2pdf4doc: bring in the SDoc RST admonitions fix
WHAT:

Brings in: https://github.com/strictdoc-project/html2pdf4doc_python/pull/84.

The most important change is a fix for StrictDoc admonition styles that were not fully covered by the page-to-page mask.

WHY:

Solves a StrictDoc issue with RST admonitions. This is the last but one known issue before we could fully enable the strict2 mode in StrictDoc's tests.